change intPair to use 64bit integers
The main use of the intPair is in counting hit/miss of the
transposition table. In a long running game the count is sure
to overflow the limits of a 32bit integer.

+++ Defines.h
@@ -38,8 +38,8 @@ enum {NOT_MATE, CHECK_MATE, STALE_MATE};
 enum {HUMAN, COMPUTER};
 
 struct intPair {
-	int one;
-	int two;
+	uint64 one;
+	uint64 two;
 };
 
 #endif
